scale_encode

Module error

An error that is emitted whenever some encoding fails.

Structs

Context

A cheaply clonable opaque context which allows us to track the current location into a type that we’re trying to encode, to aid in error reporting.

Error

An error produced while attempting to encode some type.

Location

Some location, like a field, variant or index in an array.

Enums

ErrorKind

The underlying nature of the error.

Kind

The kind of type that we’re trying to encode.

Traits

CustomError

Anything implementing this trait can be used in ErrorKind::Custom.